+++ assets/js/admin-builder-captcha.js
@@ -1,18 +1,21 @@
+/* global wpforms_builder_custom_captcha, wpforms_builder */
+
 /**
  * WPForms Custom Captcha admin builder function.
  *
  * @since 1.1.0
-*/
-var WPFormsCaptcha = window.WPFormsCaptcha || ( function( document, window, $ ) {
+ */
 
-	'use strict';
+'use strict';
+
+var WPFormsCaptcha = window.WPFormsCaptcha || ( function( document, window, $ ) {
 
 	/**
 	 * Public functions and properties.
 	 *
 	 * @since 1.1.0
 	 *
-	 * @type object
+	 * @type {object}
 	 */
 	var app = {
 
@@ -33,85 +36,177 @@ var WPFormsCaptcha = window.WPFormsCaptcha || ( function( document, window, $ ) 
 		 */
 		ready: function() {
 
-			// Cache builder element.
-			var $builder = $( '#wpforms-builder') ;
+			$( '#wpforms-builder' )
 
-			// Type (format) toggle.
-			$builder.on( 'change', '.wpforms-field-option-captcha .wpforms-field-option-row-format select', function() {
+				// Type (format) toggle.
+				.on( 'change', '.wpforms-field-option-captcha .wpforms-field-option-row-format select', app.formatToggle )
 
-				var $this = $( this ),
-					value = $this.val(),
-					id    = $this.parent().data( 'field-id' );
-				if ( value === 'math') {
-					$( '#wpforms-field-option-row-' + id + '-questions' ).hide().addClass( 'wpforms-hidden' );
-					$( '#wpforms-field-option-row-' + id + '-size' ).hide();
-				} else {
-					$( '#wpforms-field-option-row-' + id + '-questions' ).show().removeClass( 'wpforms-hidden' );
-					$( '#wpforms-field-option-row-' + id + '-size' ).show();
-				}
-			});
+				// Add new captcha question.
+				.on( 'click', '.wpforms-field-option-row-questions .add', app.addQuestion )
 
-			// Add new captcha question.
-			$builder.on( 'click', '.wpforms-field-option-row-questions .add', function( event ) {
+				// Remove captcha question.
+				.on( 'click', '.wpforms-field-option-row-questions .remove', app.removeQuestion )
 
-				event.preventDefault();
+				// Captcha sample question update.
+				.on( 'input', '.wpforms-field-option-row-questions .question', _.debounce( app.updateQuestion, 300 ) );
+		},
 
-				var $this     = $( this ),
-					$parent   = $this.parent(),
-					fieldID   = $this.closest( '.wpforms-field-option-row-questions' ).data( 'field-id' ),
-					id        = $parent.parent().attr( 'data-next-id' ),
-					$question  = $parent.clone().insertAfter( $parent );
+		/**
+		 * Format toggle event handler.
+		 *
+		 * @since 1.3.2
+		 */
+		formatToggle: function() {
 
-				$question.attr( 'data-key', id );
-				$question.find( 'input.question' ).val( '' ).attr( 'name', 'fields['+fieldID+'][questions]['+id+'][question]' );
-				$question.find( 'input.answer' ).val( '' ).attr( 'name', 'fields['+fieldID+'][questions]['+id+'][answer]' );
-				id++;
-				$parent.parent().attr( 'data-next-id', id );
-			});
+			var $this      = $( this ),
+				value      = $this.val(),
+				id         = $this.parent().data( 'field-id' ),
+				$questions = $( '#wpforms-field-option-row-' + id + '-questions' ),
+				$size      = $( '#wpforms-field-option-row-' + id + '-size' );
 
-			// Remove captcha question.
-			$builder.on( 'click', '.wpforms-field-option-row-questions .remove', function( event ) {
+			if ( value === 'math' ) {
+				$questions.hide().addClass( 'wpforms-hidden' );
+				$size.hide();
+			} else {
+				$questions.show().removeClass( 'wpforms-hidden' );
+				$size.show();
+			}
+		},
 
-				event.preventDefault();
+		/**
+		 * Add new captcha question event handler.
+		 *
+		 * @since 1.3.2
+		 *
+		 * @param {object} event Event object.
+		 */
+		addQuestion: function( event ) {
 
-				var $this = $( this ),
-					$list = $this.parent().parent(),
-					total = $list.find( 'li' ).length;
+			event.preventDefault();
 
-				if ( total === 1 ) {
-					$.alert({
-						title:   false,
-						content: wpforms_builder.error_choice,
-						icon:   'fa fa-exclamation-circle',
-						type:   'orange',
-						buttons: {
-							confirm: {
-								text:      wpforms_builder.ok,
-								btnClass: 'btn-confirm',
-								keys:     [ 'enter' ]
-							}
-						}
-					});
-				} else {
-					$this.parent().remove();
-				}
-			});
+			var $this        = $( this ),
+				$choice      = $this.closest( 'li' ),
+				$choicesList = $this.closest( '.choices-list' ),
+				fieldID      = $choicesList.data( 'field-id' ),
+				id           = $choicesList.attr( 'data-next-id' ),
+				$question    = $choice.clone().insertAfter( $choice ),
+				name         = 'fields[' + fieldID + '][questions][' + id + ']';
 
-			// Captch questions sample question
-			$builder.on( 'input', '.wpforms-field-option-row-questions li:first-of-type .question', function() {
+			$question.attr( 'data-key', id );
+			$question.find( 'input.question' ).val( '' ).attr( 'name', name + '[question]' );
+			$question.find( 'input.answer' ).val( '' ).attr( 'name', name + '[answer]' );
+			$choicesList.attr( 'data-next-id', ++id );
+		},
 
-				var $this   = $ ( this ),
-					fieldID = $this.parent().parent().data( 'field-id' );
+		/**
+		 * Remove captcha question event handler.
+		 *
+		 * @since 1.3.2
+		 *
+		 * @param {object} event Event object.
+		 */
+		removeQuestion: function( event ) {
 
-				$( '#wpforms-field-'+fieldID ).find( '.wpforms-question' ).text( $this.val() );
-			});
-		}
+			event.preventDefault();
+
+			var $this        = $( this ),
+				$choice      = $this.closest( 'li' ),
+				$choicesList = $this.closest( '.choices-list' ),
+				$questions   = $choicesList.find( '.question' ),
+				fieldID      = $choicesList.data( 'field-id' ),
+				total        = app.getTotalNotEmptyQuestions( $questions );
+
+			// We can delete a choice if at least one non-empty question will remain.
+			if (
+				total > 1 ||
+				( total === 1 && $choice.find( '.question' ).val().trim().length === 0 )
+			) {
+				$choice.remove();
+				$( '#wpforms-field-' + fieldID ).find( '.wpforms-question' ).text( $choicesList.find( '.question' ).val() );
+
+				return;
+			}
+
+			app.showAlert( null );
+		},
+
+		/**
+		 * Captcha sample question update event handler.
+		 *
+		 * @since 1.3.2
+		 */
+		updateQuestion: function() {
+
+			var $this        = $( this ),
+				$choicesList = $this.closest( '.choices-list' ),
+				$questions   = $choicesList.find( '.question' ),
+				fieldID      = $choicesList.data( 'field-id' ),
+				total        = app.getTotalNotEmptyQuestions( $questions ),
+				value        = $this.val().trim(),
+				prevValue    = $this.data( 'prev-value' ) || '';
+
+			if ( ! value.length && total < 1 ) {
+				app.showAlert( function() {
+					$this.val( prevValue );
+				} );
+
+				return;
+			}
+
+			$this.data( 'prev-value', value );
+
+			if ( $this.is( $questions[0] ) ) {
+				$( '#wpforms-field-' + fieldID ).find( '.wpforms-question' ).text( value );
+			}
+		},
+
+		/**
+		 * Show alert notifying that at least one not empty choice should remains.
+		 *
+		 * @since 1.3.2
+		 *
+		 * @param {Function|null} action Callback action attached to the confirm button.
+		 */
+		showAlert: function( action ) {
+
+			$.alert( {
+				title:   false,
+				content: wpforms_builder_custom_captcha.error_not_empty_question,
+				icon:   'fa fa-exclamation-circle',
+				type:   'orange',
+				buttons: {
+					confirm: {
+						text:     wpforms_builder.ok,
+						btnClass: 'btn-confirm',
+						keys:     [ 'enter' ],
+						action:   action,
+					},
+				},
+			} );
+		},
+
+		/**
+		 * Show alert notifying that at least one not empty choice should remains.
+		 *
+		 * @since 1.3.2
+		 *
+		 * @param {jQuery} $questions Questions choices jQuery object.
+		 *
+		 * @returns {number} Number of total not empty questions.
+		 */
+		getTotalNotEmptyQuestions: function( $questions ) {
+
+			return $questions.filter( function() {
+
+				return this.value.trim().length;
+			} ).length;
+		},
 	};
 
 	// Provide access to public functions/properties.
 	return app;
 
-})( document, window, jQuery );
+}( document, window, jQuery ) );
 
 // Initialize.
 WPFormsCaptcha.init();